ocr/pyproject.toml
[project]
name = "ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.41.2",
    "click>=8.3.1",
    "opencv-python>=4.12.0.88",
    "pydantic>=2.12.4",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
]
